bots: Move test/common/tap.py to bots/
bots/tap.py was previously a link to test/common/tap.py. Switch the
link, so that external projects can use tap.py from bots/.
